refactor(review): extract recommendation candidates
